Log sporadic request errors as Warns
A SPOT request can fail sometimes, and it's not a
real problem that should e.g be reported over
SmtpAppender if it doesn't happen too often.

So log it as Error rather than Warn if the number
of consequent request errors is less than the
threshold that set in web.config of the service.